requirements.txt
Django==3.1.2
django-extensions==3.0.9
django-import-export==2.4.0
djangorestframework==3.12.1
django-grappelli==2.14.2
